backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
duckduckgo_search==8.1.1
gunicorn==21.2.0
idna==3.11
lxml==6.0.2
packaging==25.0
Pillow==10.2.0
primp==0.15.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
redis==5.0.1
openai==1.3.0  # DeepSeek uses OpenAI-compatible API
